+++ src/HotelPosSystem/HotelDbContext.cs
@@ -12,15 +12,7 @@ namespace HotelPosSystem {
         public DbSet<Room> Rooms { get; set; }
         public DbSet<RoomType> RoomTypes { get; set; }
 
-        public string DbPath { get; }
-
-        public HotelDbContext() {
-            Environment.SpecialFolder folder = Environment.SpecialFolder.LocalApplicationData;
-            string path = Environment.GetFolderPath(folder);
-            DbPath = Path.Join(path, "hotelPosSystem.db");
-        }
-
         protected override void OnConfiguring(DbContextOptionsBuilder options)
-            => options.UseSqlite($"Data Source={DbPath}");
+            => options.UseSqlite($"Data Source={Program.DatabasePath}");
     }
 }