+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Occupied room counter for twin room 
+
 ## 0.1.0 - 2023-12-07
 
 ### Added